Adding functionality: ERB comment tokens.
ERB is Ruby embedded in HTML.
But it's also a DSL in its own right and it deserves its own commenting token; one that allows the user to use CMD+/ to easily comment out both the selected HTML •and• Ruby.

If this were to be implemented in text editor defaults it would make certain workflows much easier:
e.g. sketching out a view before the underlying model is complete to aid in considerations for model development.

This addition should not present any trouble for old implementations though it'd be great if there is a good existing test suite to confirm this.